[refactoring] Service, Activity 위치 변경
