+++ app/src/main/java/com/mashup/dorabangs/DoraShareActivity.kt
@@ -1,4 +1,4 @@
-package com.dorabangs.share
+package com.mashup.dorabangs
 
 import android.content.Intent
 import android.os.Bundle
